+++ receivers/schema/schema.go
@@ -1,8 +1,6 @@
 package schema
 
 import (
-	"errors"
-	"fmt"
 	"strings"
 )
 
@@ -208,53 +206,6 @@ func InitSchema(s IntegrationTypeSchema) IntegrationTypeSchema {
 		s.Versions[i].typeSchema = &s
 	}
 	return s
-}
-
-func ValidateTypeSchema(schema IntegrationTypeSchema) error {
-	if schema.Type == "" {
-		return errors.New("type is required")
-	}
-	if schema.Name == "" {
-		return errors.New("name is required")
-	}
-	if schema.CurrentVersion == "" {
-		return errors.New("current version is required")
-	}
-	if len(schema.Versions) == 0 {
-		return errors.New("at least one version is required")
-	}
-	found := false
-	for i := range schema.Versions {
-		if err := ValidateSchemaVersion(schema.Versions[i]); err != nil {
-			return fmt.Errorf("invalid version [%d]: %w", i, err)
-		}
-		if schema.Versions[i].Version == schema.CurrentVersion {
-			found = true
-			break
-		}
-	}
-	if !found {
-		return errors.New("current version not found")
-	}
-	return nil
-}
-
-func ValidateSchemaVersion(version IntegrationSchemaVersion) error {
-	if version.Version == "" {
-		return errors.New("version is required")
-	}
-	if len(version.Options) == 0 {
-		return errors.New("at least one option is required")
-	}
-	if version.typeSchema == nil {
-		return errors.New("type schema is not assigned")
-	}
-	for _, o := range version.Options {
-		if o.PropertyName == "" {
-			return errors.New("property name is required")
-		}
-	}
-	return nil
 }
 
 type IntegrationFieldPath []string
